ctf_task/requirements.txt
Flask
pycryptodome
python-dotenv